+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## To be announcted in 2.1
+## To be announced in 2.1
 
 - Widgets processing moved to the data warehouses for table and query sources
 - Developer API change:
@@ -9,12 +9,16 @@
 
 ## Not released
 
+-
+
+## 2.0
+
+### 2.0.6 (2023-06-07)
+
 - Feature selection UI refactor [#697](https://github.com/CartoDB/carto-react/pull/697)
 - Fix Theme overrides [#696](https://github.com/CartoDB/carto-react/pull/696)
 - Add Tooltip to Mui TablePagination and TableWidgetUI [#695](https://github.com/CartoDB/carto-react/pull/695)
 - Add typed imports, from deck.gl typed [#693](https://github.com/CartoDB/carto-react/pull/693)
-
-## 2.0
 
 ### 2.0.5 (2023-05-26)